My attempt at a slow sorting algorithm
I wanted to set the restriction on myself that the algorithm shouldn't be slow by trivial means, like a random number generator that determines whether elements are swapped. Instead, I devised a method that requires a minimum of n * (n+1) comparisons to be made. This number will become somewhat larger if duplicate elements are present, but will mostly be fairly constant. For the moment it only sorts vectors of integers, because I'm not totally sure how I would do it (passing in arrays was proving difficult, and c++ won't return auto). Anyways, I'm sure you get the picture